+++ CHANGELOG.md
@@ -28,5 +28,16 @@
 - 01 Jul 2023:
   - [feature] Added a function `supy.util.get_spinup_state` to retrieve the spin-up state for the model, which can be used for debugging and initialising the model for simulation.
 
+- 18 Oct 2023:
+  - [change] `Snow` is temporarily turned off for easier implementation of other functionalities; will be brought back in the future.
+
 - 17 May 2024:
   - [maintenance] Changed the python build backend to `meson` and `ninja` for faster builds (#257)
+
+- 31 May 2024:
+  - [feature] Added `dict_debug` an optional output of `run_supy` to help debug the model (for developers: add a `debug` flag to `df_state` to activate this feature) (#233)
+
+- 04 Jul 2024:
+  - [bugfix] Fixed a bug causing an abrupt change in results due to a less smooth transition in `z0` from surfaces without roughness elements to those with them. (#271)
+  - [bugfix] Improved the discretisation of the vertical levels in the RSL scheme for better interpolation of surface diagnostics (e.g. `T2`) (#271)
+  - [maintenance] Added support for NumPy 2.0 (#271)